created seperate release component and replaced it in RecentReleases.tsx
