+++ frontend/src/components/RecentReleases.tsx
@@ -1,18 +1,11 @@
-import { faCalendar, faFolderOpen, faTag } from '@fortawesome/free-solid-svg-icons'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { Tooltip } from '@heroui/tooltip'
-import Image from 'next/image'
-import Link from 'next/link'
-import { useRouter } from 'next/navigation'
+import { faTag } from '@fortawesome/free-solid-svg-icons'
 import React from 'react'
-import type { Release } from 'types/release'
-import { formatDate } from 'utils/dateFormatter'
+import type { Release as ReleaseType } from 'types/release' // renamed it because of the name conflict with the Release component
 import AnchorTitle from 'components/AnchorTitle'
+import Release from 'components/Release'
 import SecondaryCard from 'components/SecondaryCard'
-import { TruncatedText } from 'components/TruncatedText'
-
 interface RecentReleasesProps {
-  data: Release[]
+  data: ReleaseType[]
   showAvatar?: boolean
   showSingleColumn?: boolean
 }
@@ -22,8 +15,6 @@ const RecentReleases: React.FC<RecentReleasesProps> = ({
   showAvatar = true,
   showSingleColumn = false,
 }) => {
-  const router = useRouter()
-
   return (
     <SecondaryCard
       icon={faTag}
@@ -38,66 +29,12 @@ const RecentReleases: React.FC<RecentReleasesProps> = ({
           className={`grid ${showSingleColumn ? 'grid-cols-1' : 'gap-4 gap-y-0 sm:grid-cols-1 md:grid-cols-2 lg:grid-cols-3'}`}
         >
           {data.map((item, index) => (
-            <div key={index} className="mb-4 w-full rounded-lg bg-gray-200 p-4 dark:bg-gray-700">
-              <div className="flex w-full flex-col justify-between">
-                <div className="flex w-full items-center">
-                  {showAvatar && item?.author && (
-                    <Tooltip
-                      closeDelay={100}
-                      content={item.author.name || item.author.login}
-                      id={`avatar-tooltip-${index}`}
-                      delay={100}
-                      placement="bottom"
-                      showArrow
-                    >
-                      <Link
-                        className="flex-shrink-0 text-blue-400 hover:underline"
-                        href={item.author.login ? `/members/${item.author.login}` : '#'}
-                      >
-                        <Image
-                          alt={item.author.name || item.author.login}
-                          className="mr-2 h-6 w-6 rounded-full"
-                          height={24}
-                          src={item.author.avatarUrl}
-                          width={24}
-                        />
-                      </Link>
-                    </Tooltip>
-                  )}
-                  <h3 className="flex-1 overflow-hidden text-ellipsis whitespace-nowrap font-semibold">
-                    <Link
-                      className="text-blue-400 hover:underline"
-                      href={`https://github.com/${item.organizationName}/${item.repositoryName}/releases/tag/${item.tagName}`}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                    >
-                      <TruncatedText text={item.name || item.tagName} />
-                    </Link>
-                  </h3>
-                </div>
-                <div className="mt-2 flex flex-wrap items-center text-sm text-gray-600 dark:text-gray-400">
-                  <div className="mr-4 flex items-center">
-                    <FontAwesomeIcon icon={faCalendar} className="mr-2 h-4 w-4" />
-                    <span>{formatDate(item.publishedAt)}</span>
-                  </div>
-                  <div className="flex flex-1 items-center overflow-hidden">
-                    <FontAwesomeIcon icon={faFolderOpen} className="mr-2 h-5 w-4" />
-                    <button
-                      className="cursor-pointer overflow-hidden text-ellipsis whitespace-nowrap text-gray-600 hover:underline dark:text-gray-400"
-                      disabled={!item.organizationName || !item.repositoryName}
-                      onClick={() => {
-                        const org = item.organizationName || ''
-                        const repo = item.repositoryName || ''
-                        if (!org || !repo) return
-                        router.push(`/organizations/${org}/repositories/${repo}`)
-                      }}
-                    >
-                      <TruncatedText text={item.repositoryName} />
-                    </button>
-                  </div>
-                </div>
-              </div>
-            </div>
+            <Release
+              key={`${item.organizationName}/${item.repositoryName}@${item.tagName ?? index}`}
+              showAvatar={showAvatar}
+              item={item}
+              index={index}
+            />
           ))}
         </div>
       ) : (
